+++ tests/Unit/AllFilesTest.php
@@ -2,29 +2,21 @@
 
 namespace Sfneal\Helpers\Aws\S3\Tests\Unit;
 
+use Illuminate\Support\Facades\Storage;
 use Sfneal\Helpers\Aws\S3\StorageS3;
 use Sfneal\Helpers\Aws\S3\Tests\TestCase;
 
 class AllFilesTest extends TestCase
 {
-    /** @test */
-    public function all_files_can_be_found()
+    private function executeAssertions($expected, $allFiles)
     {
-        $expected = [
-            'article.pdf',
-            'charts.pdf',
-            'elevation.jpg',
-            'floor-plan.png',
-            'manual.pdf',
-            'workbook.pdf',
-            'directory/2021/20210002_second/article.pdf',
-            'directory/2021/20210002_second/charts.pdf',
-            'directory/2021/20210002_second/elevation.jpg',
-            'directory/2021/20210002_second/floor-plan.png',
-            'directory/2021/20210002_second/manual.pdf',
-            'directory/2021/20210002_second/workbook.pdf',
-        ];
-        $allFiles = StorageS3::key('/')->allFiles();
+        $this->assertNotNull($allFiles);
+        $this->assertIsArray($allFiles);
+
+        foreach ($expected as $expect) {
+            $this->assertTrue(in_array($expect, $allFiles));
+            Storage::disk(config('filesystem.cloud', 's3'))->exists($expect);
+        }
 
         sort($expected);
         sort($allFiles);
@@ -32,20 +24,40 @@ class AllFilesTest extends TestCase
     }
 
     /** @test */
+    public function all_files_can_be_found()
+    {
+        $this->executeAssertions(
+            [
+                'article.pdf',
+                'charts.pdf',
+                'elevation.jpg',
+                'floor-plan.png',
+                'manual.pdf',
+                'workbook.pdf',
+                'directory/2021/20210002_second/article.pdf',
+                'directory/2021/20210002_second/charts.pdf',
+                'directory/2021/20210002_second/elevation.jpg',
+                'directory/2021/20210002_second/floor-plan.png',
+                'directory/2021/20210002_second/manual.pdf',
+                'directory/2021/20210002_second/workbook.pdf',
+            ],
+            StorageS3::key('/')->allFiles()
+        );
+    }
+
+    /** @test */
     public function all_files_in_directory_can_be_found()
     {
-        $expected = [
-            'directory/2021/20210002_second/article.pdf',
-            'directory/2021/20210002_second/charts.pdf',
-            'directory/2021/20210002_second/elevation.jpg',
-            'directory/2021/20210002_second/floor-plan.png',
-            'directory/2021/20210002_second/manual.pdf',
-            'directory/2021/20210002_second/workbook.pdf',
-        ];
-        $allFiles = StorageS3::key('directory/2021/20210002_second')->allFiles();
-
-        sort($expected);
-        sort($allFiles);
-        $this->assertEquals($expected, $allFiles);
+        $this->executeAssertions(
+            [
+                'directory/2021/20210002_second/article.pdf',
+                'directory/2021/20210002_second/charts.pdf',
+                'directory/2021/20210002_second/elevation.jpg',
+                'directory/2021/20210002_second/floor-plan.png',
+                'directory/2021/20210002_second/manual.pdf',
+                'directory/2021/20210002_second/workbook.pdf',
+            ],
+            StorageS3::key('directory/2021/20210002_second')->allFiles()
+        );
     }
 }